Count Fridays the 13-th
Problem Statement

Some superstitious people believe that Friday the 13th is a day that is
very unlucky.

However, Sergey does not agree with them. In order to prove this to
people that nothing scary will happen, he wants to choose two dates and
count the number of Friday the 13th's between them. If this number is
reasonably large, he will be able to tell that it is just a regular day.

But now Sergey is quite busy, so he is unable to calculate the number
of Fridays the 13th's between the specific dates. Could you write a
program that will do it for him?

Input Format

The first line of input consists of a single integer T denoting the
number of test cases. Then, there are T lines describing the test cases.

The first (and only) line of the test case contain six single
space-separated integers D1 M1 Y1 D2 M2 Y2, denoting two dates in the
DD MM YYYY format both. It is guaranteed that they describe two
existing dates and the first date is not later than the second one.

Constraints
1≤T≤10
1900≤Y1≤Y2≤9999
D1,M1,Y1 describing a correct date
D2,M2,Y2 describing a correct date
The date, described by D1,M1,Y1 is not later than the date described by
D2,M2,Y2.

Output Format

For each specified range of dates output the number of Fridays the
13th's on a separate line.

Sample Input

2
01 10 2000 20 10 2000
19 10 1995 11 07 2015

Sample Output

1
34

Explanation

    Sample case #1: There was a Friday the 13th in October of the year
2000.
    Sample case #2: There were 34 Friday the 13th's between the
problemsetter's birthday and the day he set this problem.